s3_sync - fix when running in FIPS mode
